refactor(types): utility helpers
